% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzuko.R
\name{batch_normalization}
\alias{batch_normalization}
\title{Batch Normalization for re-scaling a set of units}
\usage{
batch_normalization(
  x,
  gamma = NULL,
  beta = NULL,
  mean_x = NULL,
  var_x = NULL,
  epsilon = 1e-08
)
}
\arguments{
\item{x}{The matrix of units to be batch-normalized}

\item{gamma}{A vector of re-scaling parameters}

\item{beta}{A vector of re-shifting constants}

\item{epsilon}{The additive constant for numerical stability}
}
\value{
The batch-normalized matrix of re-scaled units
}
\description{
Batch Normalization for re-scaling a set of units
}
\note{
For Internal Use. The input matrix should have no intercept.
}
