% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzuko.R
\name{cross_entropy}
\alias{cross_entropy}
\title{Compute the negative cross-entropy for multi-class classification}
\usage{
cross_entropy(p, y)
}
\arguments{
\item{p}{The predictions}

\item{y}{The outputs}
}
\value{
The negative cross-entropy
}
\description{
Compute the negative cross-entropy for multi-class classification
}
\note{
For Internal Use
}
