% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzuko.R
\name{forward_backward_pass}
\alias{forward_backward_pass}
\title{Compute crucial quantities evaluated from one forward-Backward pass through the neural network}
\usage{
forward_backward_pass(
  x,
  y,
  w,
  activation,
  output_type,
  dropout = FALSE,
  retain_rate = NULL,
  forward_only = FALSE
)
}
\arguments{
\item{x}{The inputs}

\item{y}{The outputs}

\item{w}{The list of weights: 1st element are connection of weights from input to 1st hidden layer,
and the last element are connection weights from the last hidden layer to the outputs}

\item{output_type}{The output type: either "numeric" (regression) or "categorical" (prediction).}

\item{dropout}{Boolean to indicate whether dropout is used}

\item{retain_rate}{The proportion of units to retain during dropout}

\item{forward_only}{If TRUE the function will only evaluate the forward pass}
}
\value{
A list containing the following elements:
p: the output probabilities
delta: a list of errors backpropagated throught the layers
z: the hidden units values
}
\description{
Compute crucial quantities evaluated from one forward-Backward pass through the neural network
}
