% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzuko.R
\name{forward_pass}
\alias{forward_pass}
\title{Compute one forward pass through the neural network}
\usage{
forward_pass(x, y, w, activation, output_type)
}
\arguments{
\item{x}{The inputs}

\item{y}{The outputs}

\item{w}{The list of weights: 1st element are connection of weights from input to 1st hidden layer,
and the last element are connection weights from the last hidden layer to the outputs}

\item{output_type}{The output type: either "numeric" (regression) or "categorical" (prediction).}
}
\value{
A list with the following elements:
p: the output probabilities
z: the hidden units values
}
\description{
Compute one forward pass through the neural network
}
