% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzuko.R
\name{get_error_hidden}
\alias{get_error_hidden}
\title{Compute errors from a layer to the previous layer}
\usage{
get_error_hidden(delta, grad_s, w)
}
\arguments{
\item{delta}{The error from the next layer}

\item{grad_s}{The gradient of the activation function for linear predictors at the current layer}

\item{The}{weights associated with the current layer}
}
\value{
The error term to be back-propagated
}
\description{
Compute errors from a layer to the previous layer
}
\note{
For Internal use
}
