% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzuko.R
\name{get_error_output}
\alias{get_error_output}
\title{Compute errors from output to the last hidden layer}
\usage{
get_error_output(y, p)
}
\arguments{
\item{y}{The outputs}

\item{p}{The predictions}
}
\value{
The error term (a.k.a delta) from the output to the last hidden layer
}
\description{
Compute errors from output to the last hidden layer
}
\note{
For Internal Use
}
