% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzuko.R
\name{get_s}
\alias{get_s}
\title{Compute the linear predictors to be activated via the activation function}
\usage{
get_s(x, w)
}
\arguments{
\item{w}{The weights}

\item{z}{The inputs or hidden units}
}
\value{
The linear predictors
}
\description{
Compute the linear predictors to be activated via the activation function
}
\note{
For Internal Use
}
