% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzuko.R
\name{grad_logistic}
\alias{grad_logistic}
\title{Compute the gradient of the logistic activation function}
\usage{
grad_logistic(s)
}
\arguments{
\item{s}{The linear predictors}
}
\value{
The gradient of logistic activation evaluated at s
}
\description{
Compute the gradient of the logistic activation function
}
\note{
For Internal Use
}
