% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzuko.R
\name{grad_relu}
\alias{grad_relu}
\title{Compute the gradient of the relu activation function}
\usage{
grad_relu(s)
}
\arguments{
\item{s}{The linear predictors}
}
\value{
The gradient of logistic activation evaluated at s
}
\description{
Compute the gradient of the relu activation function
}
\note{
For Internal Use. We set gradient at 0 to be 0.
}
