% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzuko.R
\name{grad_tanh}
\alias{grad_tanh}
\title{Compute the gradient of the tanh activation function}
\usage{
grad_tanh(s)
}
\arguments{
\item{s}{The linear predictors}
}
\value{
The gradient of logistic activation evaluated at s
}
\description{
Compute the gradient of the tanh activation function
}
\note{
For Internal Use
}
