% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzuko.R
\name{grad_w}
\alias{grad_w}
\title{Compute the gradient of the weight for a given layer}
\usage{
grad_w(delta, x)
}
\arguments{
\item{delta}{The errors passed from the next layer}

\item{x}{The inputs or the current hidden units}
}
\value{
The gradient of the weights for gradient descent update
}
\description{
Compute the gradient of the weight for a given layer
}
\note{
For Internal Use
}
