% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzuko.R
\name{initialize_weights}
\alias{initialize_weights}
\title{Initialize weights}
\usage{
initialize_weights(x_train, y_train, num_hidden, method)
}
\arguments{
\item{x_train}{The training inputs}

\item{y_train}{The training outputs}

\item{num_hidden}{Vector with number of hidden units for each hidden layer}

\item{method}{The initialization method}
}
\value{
Initial weights
}
\description{
Initialize weights
}
\note{
For Internal Use
}
