% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzuko.R
\name{kl_divergence}
\alias{kl_divergence}
\title{Compute the KL-divergence for logistic output}
\usage{
kl_divergence(p, y)
}
\arguments{
\item{p}{The predictions}

\item{y}{The outputs}
}
\value{
The KL divergence
}
\description{
Compute the KL-divergence for logistic output
}
\note{
For Internal Use
}
