% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzuko.R
\name{least_square}
\alias{least_square}
\title{Compute 1/2 least square error for regression}
\usage{
least_square(p, y)
}
\arguments{
\item{p}{The predictions}

\item{y}{The outputs}
}
\value{
The least square error (halved)
}
\description{
Compute 1/2 least square error for regression
}
\note{
For Internal Use. Note the least square error is halved
to simplify the gradients
}
