% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzuko.R
\name{logistic_activation}
\alias{logistic_activation}
\title{Compute logistic activation given linear predictors}
\usage{
logistic_activation(s)
}
\arguments{
\item{s}{The linear predictors}
}
\value{
The unit activations
}
\description{
Compute logistic activation given linear predictors
}
\note{
For Internal Use
}
