% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzuko.R
\name{pretrain}
\alias{pretrain}
\title{Initialize weights by Stacking Autoencoders}
\usage{
pretrain(
  x_train,
  y_train,
  num_hidden,
  iter = 300,
  step_size = 0.01,
  lambda = 1e-05,
  momentum = 0.9,
  ini_method = c("normalized", "uniform", "gaussian"),
  sparse = FALSE,
  verbose = F
)
}
\arguments{
\item{x_train}{The training inputs}

\item{y_train}{The training outputs}

\item{num_hidden}{Vector with number of hidden units for each hidden layer}

\item{method}{The initialization method}
}
\value{
Initial weights
}
\description{
Initialize weights by Stacking Autoencoders
}
\note{
For Internal Use
}
