% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzuko.R
\name{relu_activation}
\alias{relu_activation}
\title{Compute ReLU activation given linear predictors}
\usage{
relu_activation(s)
}
\arguments{
\item{s}{The linear predictors}
}
\value{
The unit activations
}
\description{
Compute ReLU activation given linear predictors
}
\note{
For Internal Use
}
