% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzuko.R
\name{soft_max}
\alias{soft_max}
\title{Compute the soft max activation for output predictive probabilities}
\usage{
soft_max(a)
}
\arguments{
\item{a}{The linear predictors from the last hidden layer}
}
\value{
The output probabilities
}
\description{
Compute the soft max activation for output predictive probabilities
}
\note{
For Internal Use
}
