% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzuko.R
\name{tanh_activation}
\alias{tanh_activation}
\title{Compute tanh activation given linear predictors}
\usage{
tanh_activation(s)
}
\arguments{
\item{s}{The linear predictors}
}
\value{
The unit activations
}
\description{
Compute tanh activation given linear predictors
}
\note{
For Internal Use
}
