% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzuko.R
\name{get_error_hidden}
\alias{get_error_hidden}
\title{Compute errors from a layer to the previous layer}
\usage{
get_error_hidden(delta, grad_s, w, batch_norm = F, batch_norm_obj = NULL)
}
\arguments{
\item{delta}{The error from the next layer}

\item{grad_s}{The gradient of the activation function for linear predictors at the current layer}

\item{The}{weights associated with the current layer}

\item{x}{The original hidden units}

\item{gamma}{The gamma parameter from batch-norm}

\item{mean_x}{The column means of x}

\item{sd_x}{The column standard deviation of x (+ epsilon)}
}
\value{
The error term to be back-propagated
}
\description{
Compute errors from a layer to the previous layer
}
\note{
For Internal use
}
