% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzuko.R
\name{get_error_output}
\alias{get_error_output}
\title{Compute errors from output to the last hidden layer}
\usage{
get_error_output(y, p, batch_norm = F, batch_norm_obj = NULL)
}
\arguments{
\item{y}{The outputs}

\item{p}{The predictions}

\item{batch_norm}{If batch_norm is applied}

\item{x}{The original hidden units}

\item{gamma}{The gamma parameter from batch-norm}

\item{mean_x}{The column means of x}

\item{sd_x}{The column standard deviation of x (+ epsilon)}
}
\value{
The error term (a.k.a delta) from the output to the last hidden layer
}
\description{
Compute errors from output to the last hidden layer
}
\note{
For Internal Use
}
