% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzuko.R
\docType{data}
\name{mnist}
\alias{mnist}
\title{MNIST Database Training Images}
\format{
A list with 2 elements
\describe{
  \item{x_train}{A matrix of 784 pixel intensites of 60000 images, normalized to range [0,1]}
  \item{y_train}{The digit of each image, from 0 to 9}
}
}
\source{
\url{http://yann.lecun.com/exdb/mnist/}
}
\usage{
mnist
}
\description{
Pixel Intensities and Class Labels of 60000 Training Images from the
MNIST database
}
\note{
The data was downloaded from the source below, and included
"train-labels-idx1-ubyte" and "train-images-idx3-ubyte". A script from
https://github.com/rstudio/tfestimators/blob/master/vignettes/examples/mnist.R
was used to decode the original file format to pixel intensities.
Only the 60000 training images are included in this file.
}
\keyword{datasets}
