% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzuko.R
\name{netzuko}
\alias{netzuko}
\title{Fit a neural network using back-propagation}
\usage{
netzuko(
  x_train,
  y_train,
  x_test,
  y_test,
  num_hidden = c(3, 3),
  iter = 10000,
  step_size = 0.5,
  lambda = 1e-05,
  momentum = 0.9,
  ini_w = NULL,
  sparse = FALSE,
  verbose = F
)
}
\arguments{
\item{x_train}{The training inputs}

\item{y_train}{The training outputs}

\item{x_test}{The test inputs}

\item{y_test}{The test outputs}

\item{num_hidden}{A vector with length equal the number of hidden layers, and
values equal the number of hidden units in the corresponding layer. The default c(2, 2) will fit
a neural network with 2 hidden layers with 2 hidden units in each layer.}

\item{iter}{The number of iterations of gradient descent}

\item{step_size}{The step size for gradient descent}

\item{lambda}{The weight decay parameter}

\item{momentum}{The momentum for the momentum term in gradient descent}

\item{ini_w}{A list of initial weights. If not provided the function will initialize the weights
automatically by simulating from a Gaussian distribution with small variance.}

\item{sparse}{If the input matrix is sparse, setting sparse to TRUE can speed up the code.}

\item{verbose}{Will display fitting progress when set to TRUE}
}
\value{
A list containing the following elements:

cost_train The training cost by iteration

cost_test: The test cost by iteration

w: The list of weights at the final iteration
}
\description{
Fit a neural network using back-propagation
}
\examples{
set.seed(8)
logistic = function(alpha, beta, x) 1/(1 + exp(-(alpha + beta*x)))
x_train = matrix(rnorm(300), 100, 3)
y_train = factor(rbinom(100, 1, prob = logistic(alpha = 0, beta = 1, x_train[,1])) +
                  rbinom(100, 1, prob = logistic(alpha = 0, beta = 1, x_train[,2])))
x_test = matrix(rnorm(3000), 1000, 3)
y_test = factor(rbinom(1000, 1, prob = logistic(alpha = 0, beta = 1, x_test[,1])) +
                 rbinom(1000, 1, prob = logistic(alpha = 0, beta = 1, x_test[,2])))
fit = netzuko(x_train, y_train, x_test, y_test, step_size = 0.1, iter = 200)
plot(fit$cost_train, type = "l")
lines(fit$cost_test, col = 2)
}
