% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netzuko.R
\name{scale_matrix}
\alias{scale_matrix}
\title{Scale a matrix}
\usage{
scale_matrix(x, mean_x = NULL, sd_x = NULL, intercept = F)
}
\arguments{
\item{x}{The matrix to be scaled}

\item{mean_x}{The means to subtract. Will be computed if NULL}

\item{sd_x}{The standard deviation to devide. Will be computed if NULL}

\item{intercept}{Indicates if the first column of x is an intercept (all 1's)}
}
\value{
A list containing the scaled matrix, and the mean
and standard deviations used by scaling
}
\description{
Scale a matrix
}
\note{
For Internal Use
}
